requirements.txt
# Core dependencies
numpy==1.26.4
torch==2.1.2
transformers==4.38.2
peft==0.8.2

# Data processing
datasets==2.18.0

# Evaluation
nltk==3.8.1
rouge_score==0.1.2
scikit-learn==1.4.1

# Multi-modal
torchvision==0.16.2
ftfy==6.1.3
regex==2023.12.25

# Vector search
faiss-cpu==1.7.4

# Visualization
matplotlib==3.8.4
seaborn==0.13.2

# Utilities
python-dotenv==1.0.1

# Knowledge base
sentence-transformers==2.5.1
